campuscopilot_backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
aiosqlite
python-dotenv
openai
requests
apscheduler